Internal Types

There are two important objects when thinking about trace rescaling. First is the storm trace, and second is its summary. We store the trace as a StormTrace and the summary as a Vector{Float64}. A history of storm traces and summaries is then stored as a StormHistory.

Storm Traces

The storm trace type contains both the time points (a range) and the values (a matrix).

HistoricalStormTraceSimulation.StormTrace — Type

StormTrace(value,time)

Construct a StormTrace type used for storing a storm type.

Arguments

value: A matrix of floats which is n × p.

time: An abstract range of length n.

You can compute the number of variables using

HistoricalStormTraceSimulation.nvariables — Function

nvariables(t::StormTrace)

Compute the number of variable contained by a storm trace, including time.

This is size(t.value,2)+1.

You can compute the length (number of time points) with

Base.length — Method

length(t::StormTrace)

Compute the length of a StormTrace, which is the number of time points (not total number of elements).

Storm History

HistoricalStormTraceSimulation.StormHistory — Type

StormHistory(summaries,traces)

For storing historical storms.

Arguments

summaries: Should be a vector of vectors each of length q, containing summaries of storms.

traces: Should be a vector of StormTraces, each satisfying nvariables(t)==q.

You can compute the number of storms with

Base.length — Method

length(s::StormHistory)

Compute the length of a StormHistory, i.e. the number of storms contained in the history.