Basic usage

The HistoricalStormTraceSimulation.jl package is designed to generate storm traces based on historical storms to match simulated storm characteristics. There are two main interface functions which should be used:

Missing docstring.

Missing docstring for sampletraces. Check Documenter's build log for details.

This is the main function used to generate traces from historical storms.

HistoricalStormTraceSimulation.dataframes2storms — Function

dataframes2storms(event_data, event_start_end, input_data, simulated_data)

Convert dataframes containing storm parameters and data to traces and summaries for use in package.

Will reorder variables to match up names of variables. Pass outputs to sampletraces function.

Arguments:

event_data - DataFrame containing summaries of historical storms.

event_start_end - DataFrame containing start and end indices of events in input_data.

input_data - DataFrame containing historical time series.

simulated_data - DataFrame containing simulated storm summaries.

Outputs:

new_summaries - Vector of summary vectors.

history - StormHistory object.

summary_names - Names of summary variables in order (traces are the same but one less variable (time is separate)).

This is a helper function to convert standard formats from other packages/languages to appropriate formats for this package. Note that it is important that naming conventions are consistent across data frames for this function to work.